pyproject.toml
[project]
name = "Ai-researcher"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Lacey W.", email = "lacey.m.washington@gmail.com"},
]
dependencies = ["arxiv>=2.2.0", "llama-index>=0.12.29", "python-dotenv>=1.1.0", "notebook>=7.4.0", "streamlit>=1.44.1", "fitz>=0.0.1.dev2", "tenacity>=9.1.2", "pypdf2>=3.0.1", "langchain>=0.3.23", "sentence-transformers>=4.1.0", "pymupdf>=1.25.5", "pytesseract>=0.3.13", "pdf2image>=1.17.0", "pillow>=11.2.1", "pdfplumber>=0.11.6", "pinecone>=6.0.2", "llama-index-vector-stores-pinecone>=0.4.5", "ipywidgets>=8.1.6", "jupyter>=1.1.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

